Show keypoints on the classes page; Add meta json download; Move Label to the annotation_widgets/image
