+++ annotation_widgets/image/filtering/widget.py
@@ -10,7 +10,6 @@ from models import ProjectData, Value
 from .gui import FilteringStatusBar
 from .io import ImageFilteringIO
 from .logic import ImageFilteringLogic
-from ..labeling.models import Label
 
 
 class ImageFilteringWidget(AbstractImageAnnotationWidget):
@@ -19,21 +18,6 @@ class ImageFilteringWidget(AbstractImageAnnotationWidget):
 
     def set_up_status_bar(self):
         self.status_bar = FilteringStatusBar(parent=self, logic=self.logic)
-
-    def show_classes(self):
-
-        data = [
-            {
-                "name": l.name,
-                "color": l.color,
-                "hotkey": l.hotkey,
-            }
-            for l in Label.get_figure_labels()
-        ]
-        env = Environment(loader=FileSystemLoader(templates_path))
-        template = env.get_template('classes.html')
-        html_content = template.render(data=data)
-        show_html_window(root=self, title="Classes", html_content=html_content)
 
     def add_menu_items(self, root: tk.Tk):
         assert root.help_menu is not None
