+++ annotation_widgets/image/labeling/bboxes/models.py
@@ -1,5 +1,6 @@
 from annotation_widgets.image.labeling.drawing import draw_text_label
-from annotation_widgets.image.labeling.models import Figure, Label, Point
+from annotation_widgets.image.labeling.models import Figure, Point
+from annotation_widgets.image.models import Label
 from db import get_session
 
 import json
