+++ annotation_widgets/image/labeling/segmentation/figure_controller.py
@@ -1,5 +1,6 @@
+from annotation_widgets.image.models import Label
 import cv2
-from annotation_widgets.image.labeling.models import Figure, Label
+from annotation_widgets.image.labeling.models import Figure
 from annotation_widgets.image.labeling.segmentation.models import Mask
 from config import ColorBGR
 from annotation_widgets.image.labeling.figure_controller import AbstractFigureController, Mode
