+++ templates/classes.html
@@ -42,12 +42,15 @@
         <tr>
           <th style="width: 100px;">Color</th>
           <th style="width: 300px;">Label</th>
+          <th style="width: 100px;">Type</th>
           <th style="width: 100px;">Hotkey</th>
         </tr>
         {% for item in data %}
         <tr>
           <td bgcolor="{{ item.color }}" height="40"></td>
           <td style="font-size: 16px;">{{ item.name }}</td>
+          <td style="font-size: 16px;">{{ item.type }}</td>
+          
           <td>{{ item.hotkey }}</td>
         </tr>
         {% endfor %}
